Improved backend

+++ demo/src/main/java/Projeto_teste/demo/service/UserService.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 
 
 import Projeto_teste.demo.repository.UserRepository;
+import Projeto_teste.demo.exceptions.UserInvalidArguments;
+import Projeto_teste.demo.exceptions.UserNotFoundException;
 import Projeto_teste.demo.model.User;
 
 import java.util.List;
@@ -19,34 +21,46 @@ public class UserService {
     UserRepository userRepository;
 
     public List<User> getAllUsers(){
-        return userRepository.findAll();
+        return userRepository.findAll().stream().toList();
     }
 
     public User getUser(Long id){
-        return userRepository.findById(id).orElse(null);
+        return userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
     }
+    
 
     public User saveUser(User user) {
-        return userRepository.save(user);
+        if(user.getName()!= null && user.getEmail() != null){
+            return userRepository.save(user);
+        }
+        else{ throw new UserInvalidArguments();}
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
     }
 
-    public User ChangeUser(Long id, String mail, String name ){
-        User user = getUser(id);
-        
-        if(mail != null){
-            user.SetEmail(mail);
-        }
-        if (name != null){
-            user.SetName(name);
+    public User changeUser(Long id, User updatedUser) {
+        User existingUser = getUser(id);
+
+        if (existingUser == null) {
+            return null; 
         }
 
-        userRepository.save(user);
+        if (updatedUser.getName() != null) {
+            existingUser.setName(updatedUser.getName());
+        }
 
-        return user; 
+
+        if (updatedUser.getEmail() != null) {
+            existingUser.setEmail(updatedUser.getEmail());
+        }
+
+        userRepository.save(existingUser);
+        return existingUser;
     }
     
 }
